feat: signal fitting
This implementation still has a bug that messes up the blanks if a single signal group contains more than one signal with the same pad.

In this case it's the UPDI and RESET signals that are both grouped in the functional group "other".

This can be worked around by refining the signal groups and splitting such duplicates into separate groups.